move the custom hooks of login and signup to src/hooks
